skills/slack_assistant/requirements.txt
# MCP Slack Assistant Dependencies

# Core MCP SDK
mcp>=1.26.0

# Slack SDK
slack-sdk>=3.23.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
